Episode-3/requirements.txt
nextcord
python-dotenv
aiosqlite